bump deps: AdguardTeam/dnsproxy@v0.83.1
- upgrade all deps: golibs v0.35.15, miekg/dns v1.1.72, quic-go
  v0.61.0, golang.org/x/* latest; go directive 1.26.5
- switch DNSCrypt to AdguardTeam/dnscrypt (project moved from
  ameshkov/dnscrypt); use Validate + NewCert API
- internal: replace adapted handler with verbatim upstream middleware;
  sync dnsmsg (NewMsgFORMERR) and netutil; adapt cmd to new proxy APIs
  (Handler.ServeDNS with ctx, proxy.HTTPConfig, ratelimit middleware,
  uint16 ports, uint ratelimit)
- add internal/update.sh: mechanized sync of the verbatim internal
  packages with upstream dnsproxy, with build verification and a
  porting guide for the adapted cmd package
- ci: actions/checkout@v7, actions/setup-go@v7, go version from
  go.mod; drop obsolete GO111MODULE
- docker: golang:1.26 base image

+++ internal/cmd/config.go
@@ -46,20 +46,20 @@ type Configuration struct {
 	// ListenAddrs is the list of server's listen addresses.
 	ListenAddrs []string `yaml:"listen-addrs"`
 
-	// ListenPorts are the ports server listens on.
-	ListenPorts []int `yaml:"listen-ports"`
+	// ListenPorts are the ports server listens on for plain DNS.
+	ListenPorts []uint16 `yaml:"listen-ports"`
 
 	// HTTPSListenPorts are the ports server listens on for DNS-over-HTTPS.
-	HTTPSListenPorts []int `yaml:"https-port"`
+	HTTPSListenPorts []uint16 `yaml:"https-port"`
 
 	// TLSListenPorts are the ports server listens on for DNS-over-TLS.
-	TLSListenPorts []int `yaml:"tls-port"`
+	TLSListenPorts []uint16 `yaml:"tls-port"`
 
 	// QUICListenPorts are the ports server listens on for DNS-over-QUIC.
-	QUICListenPorts []int `yaml:"quic-port"`
+	QUICListenPorts []uint16 `yaml:"quic-port"`
 
 	// DNSCryptListenPorts are the ports server listens on for DNSCrypt.
-	DNSCryptListenPorts []int `yaml:"dnscrypt-port"`
+	DNSCryptListenPorts []uint16 `yaml:"dnscrypt-port"`
 
 	// Upstreams is the list of DNS upstream servers.
 	Upstreams []string `yaml:"upstream"`
@@ -119,15 +119,15 @@ type Configuration struct {
 	CacheSizeBytes int `yaml:"cache-size"`
 
 	// Ratelimit is the maximum number of requests per second.
-	Ratelimit int `yaml:"ratelimit"`
+	Ratelimit uint `yaml:"ratelimit"`
 
 	// RatelimitSubnetLenIPv4 is a subnet length for IPv4 addresses used for
 	// rate limiting requests.
-	RatelimitSubnetLenIPv4 int `yaml:"ratelimit-subnet-len-ipv4"`
+	RatelimitSubnetLenIPv4 uint `yaml:"ratelimit-subnet-len-ipv4"`
 
 	// RatelimitSubnetLenIPv6 is a subnet length for IPv6 addresses used for
 	// rate limiting requests.
-	RatelimitSubnetLenIPv6 int `yaml:"ratelimit-subnet-len-ipv6"`
+	RatelimitSubnetLenIPv6 uint `yaml:"ratelimit-subnet-len-ipv6"`
 
 	// UDPBufferSize is the size of the UDP buffer in bytes.  A value <= 0 will
 	// use the system default.
